feat: renamed plugins folder
